requirements.txt
requests
beautifulsoup4
selenium
pillow
jinja2
pandas
python-slugify
tqdm
schedule
apscheduler
google-api-python-client
python-dotenv
PyYAML